add arrow mapping
